style: extract getChromeApiNamespacePaths to own util
